refactor(raft): extract ForceConfigureQuorum
